requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.5
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.7
flower==2.0.1
greenlet==3.0.1
humanize==4.8.0
kombu==5.3.4
numpy==1.26.2
pandas==2.1.3
prometheus-client==0.18.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
text-unidecode==1.3
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.10
